rc: reintroduce qcom_diag
This will avoid some misunderstood due to the new group wakelock.
If few services are in both groups wakelock and qcom_diag, this
is not the case for every ones (eg: audioserver).

Reintroduction of qcom_diag and /dev/diag group set to qcom_diag
is preferable

Add qcom_oncrpc and qcom_diag to audioserver.rc